+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # kgaut_tools 2.0.x-dev
-
+ - Create `EntityTitleTrait`
  - Add submodule `kgaut_tools_paragraphs`
  - Create a view pager PagerFullWithSpecificFirstPage
  - Clean start for 2.0 version